fix: thew new way
