+++ apps/web/components/settings/general-settings/View.vue
@@ -1,0 +1,7 @@
+<template>
+  <SiteConfigurationView>
+    <template #setting-group-title> General Settings </template>
+  </SiteConfigurationView>
+</template>
+
+<script setup lang="ts"></script>
