+++ apps/web/components/settings/general-settings/bundle-settings/types.ts
@@ -1,0 +1,1 @@
+export type SortingOption = { label: string; value: string };
